unit type fix;
light selection color fix;

+++ FireRender.Max.Plugin/plugin/light/IESprocessor.h
@@ -113,7 +113,7 @@ public:
 	* The initial rated lumens for the lamp used in the test or -1 if absolute photometry is used
 	* and the intensity values do not depend on different lamp ratings. 
 	*/
-	int m_lumens = 0;
+	double m_lumens = 0;
 
 	/**
 	* A multiplying factor for all the candela values in the file. 
@@ -122,7 +122,7 @@ public:
 	* when you obtain the photometric values from a catalog using a ruler on a goniometric diagram.
 	* Normally the multiplying factor is 1.
 	*/
-	int m_multiplier = 0;
+	double m_multiplier = 0;
 
 	/**
 	* The number of vertical (polar) angles in the photometric web.
